added: repo for user + some base routes

+++ models/quest.go
@@ -1,7 +1,7 @@
 package models
 
 type Quest struct {
-	ID   int    `json:"Id"`
+	ID   int64  `json:"Id"`
 	Name string `json:"name"`
 	Cost int    `json:"cost"`
 }
